requirements.txt
fastapi
uvicorn[standard]
gunicorn
dnspython
motor
python-jose[cryptography]
passlib[bcrypt]
jinja2
python-multipart
authlib
httpx

#graphs plotting
matplotlib

# Additional dependencies for data
pandas[parquet, feather, spss, excel]
openai
pinecone-client
tiktoken

# pypdf

#blob
azure-storage-blob 
azure-identity

# markdown
markdown

# aiohttp
aiohttp

# AI
groq

# Email
sendgrid
